requirements.txt
aiogram==3.3.0
httpx==0.27.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
structlog==23.2.0
Pillow==10.1.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
pytest
pytest-asyncio

# Optional dependencies for voice transcription and additional formats
# openai-whisper==20231117; sys_platform != 'win32'
# ffmpeg-python==0.2.0; sys_platform != 'win32'